+++ src/api/chia/farmer/farmer.ts
@@ -1,4 +1,4 @@
-import {float, int, str, uint64} from "../types/_python_types_";
+import {float, int, Optional, str, uint64, uint8} from "../types/_python_types_";
 import {TPartialsErrorResponse, TPoolInfoResponse} from "../../rpc/pool/index";
 import {PoolWalletConfig} from "../pools/pool_config";
 
@@ -32,10 +32,12 @@ export type PoolState = {
   points_found_24h: Array<[float, uint64]>;
   points_acknowledged_since_start: int;
   points_acknowledged_24h: Array<[float, uint64]>;
-  current_points_balance: int;
-  current_difficulty: int;
+  next_farmer_update: float;
+  next_pool_info_update: float;
+  current_points: uint64;
+  current_difficulty: uint64;
   pool_errors_24h: TPartialsErrorResponse[];
-  pool_info: TPoolInfoResponse;
+  authentication_token_timeout: uint8;
   pool_config: PoolWalletConfig;
   p2_singleton_puzzle_hash: str;
 };