+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Update anchor
 https://github.com/Chia-Network/chia-blockchain/compare/cd78b0e99f199ed97a357775234ae2de0f7eb8b3...pools.2021-may-25
 
+## [1.1.0-beta.9] (Changes from 1.1.0-beta.8)
+### Changed
+- Updated `PoolState` type.
+- Changed API endpoint name `set_pool_payout_instructions` to `set_payout_instructions`
+
+### Added
+- Added type TCreate_New_Pool_WalletRequest
+- Added type TCreate_New_Pool_WalletResponse
+
 ## [1.1.0-beta.8] (Changes from 1.1.0-beta.7)
 ### Added
 - Added new full_node API [`get_puzzle_and_solution`](./src/api/rpc/full_node/README.md#get_puzzle_and_solutionagent-params)
@@ -36,7 +45,7 @@ https://github.com/Chia-Network/chia-blockchain/compare/cd78b0e99f199ed97a357775
   - [`login`](./src/api/rpc/pool/README.md#loginagent-params)
 - [New farmer api](./src/api/rpc/farmer)
   - [`get_pool_state`](./src/api/rpc/farmer/README.md#get_pool_stateagent)
-  - [`set_pool_payout_instructions`](./src/api/rpc/farmer/README.md#set_pool_payout_instructionsagent-params)
+  - [`set_payout_instructions`](./src/api/rpc/farmer/README.md#set_payout_instructionsagent-params)
   - [`get_plots`](./src/api/rpc/farmer/README.md#get_plotsagent)
 - [New full_node api](./src/api/rpc/full_node)
   - [`get_recent_signage_point_or_eos`](./src/api/rpc/full_node/README.md#get_recent_signage_point_or_eosagent-params)
@@ -107,7 +116,8 @@ daemon.sendMessage(destination, get_block_record_by_height_command, data);
 Initial release.
 
 <!-- [Unreleased]: https://github.com/Chia-Mine/chia-agent/compare/v0.0.1...v0.0.2 -->
-[1.1.0-beta.8]: https://github.com/Chia-Mine/chia-agent/compare/2aea73f07b40f0a256e55805b84d189a6c67ea3b...v1.1.0
+[1.1.0-beta.9]: https://github.com/Chia-Mine/chia-agent/compare/16ef8414950c4d0d34d33d9d6307a78654df59dd...v1.1.0
+[1.1.0-beta.8]: https://github.com/Chia-Mine/chia-agent/compare/2aea73f07b40f0a256e55805b84d189a6c67ea3b...16ef8414950c4d0d34d33d9d6307a78654df59dd
 [1.1.0-beta.7]: https://github.com/Chia-Mine/chia-agent/compare/1d5065fcd0b1b673fd7c7e042205ca34bbc66836...2aea73f07b40f0a256e55805b84d189a6c67ea3b
 [1.1.0-beta.6]: https://github.com/Chia-Mine/chia-agent/compare/18eb244e4adef48b7c8a6ed70719230fc237f6b4...1d5065fcd0b1b673fd7c7e042205ca34bbc66836
 [1.1.0-beta.5]: https://github.com/Chia-Mine/chia-agent/compare/97bd55fe016f52c4a5fce313bd1e7cb8a26fa7e0...18eb244e4adef48b7c8a6ed70719230fc237f6b4